requirements.txt
gql>=3.4.0
requests>=2.28.0
python-dotenv>=0.20.0
dataclasses>=0.6
typing-extensions>=4.3.0
